Log started message after socket file has been setup
There seems to be some weirdness a significant delay between the
time signald is started and the time the socket file appears. This
change should cause the initial "started" message to appear after
the socket file is ready. A debug-level "starting" message is logged
where the old info-level started message was.