Implement subscribe and unsubscribe v1
Refactors how the client connection loop works to remove dependence
on the legacy socket handler. In testing, existing functionality
seems mostly unaffected. Open an issue if I broke you.

the MessageReceiver's subscribe() method now takes a MessageEncoder
as well as an account name and socket. The MessageEncoder encodes
messages and write's them to the socket. Although the two currently
implemented MessageEncoders produce JSON, the interface provides
them raw socket access so other things could be used in the future
(like protobufs see #107). The v0 implementation mostly uses existing
code, although some of it had to be duplicated into clientprotocol/v0
to eliminate cross version referencing in the incoming protocol.

The v1 subscribe protocol similar to the v0 protocol, but not identical.
The wrapper format has been changed to the structure described in
the v1 type ClientMessageWrapper. Values are identified by their
version and type name. Errors are reported in the same structure,
but have a boolean field to indicate they are an error.

There have been some changes to the errors in the v1 protocol in
preperation for adding them to the protocol document (#149)

Now that all v0 functionality is availabe in v1, v0 requests are
deprecated. v0 support will remain in signald releases through the
end of 2021, after which they will be removed when I get around to
it. This information is encoded in an annotation now, which is added
to the protocol document under the "removal_date" field.

All of these changes have resulted in a lot of new files, many of which
are nearly identical to existing files duplicated to reduce parts
of the versioned client protocol that can't easily be changed in
one version without changing it in the other(s).

closes #88