server is now a per-account property.
Servers are tracked in the db and can be manipulated by clients. Two
servers are pre-populated: Signal's production and staging. You can
add your own with the new add_server request type. The server may
be specified by UUID when linking or registering. If not specified
the default will be used. Default is configurable at build time, if
unchanged Signal's production servers will be uesd.

fixes #188