Fix handling of edge case address resolution
if a UUID is in the database with not e164 and we try to send a message
to the e164 associated with that UUID, the resolution code would find
the UUID, not realize it was already known and try to insert a new row,
which violates the UUID columns uniqueness constraint and causes a crash.

This fixes by re-running the entire get() sequence once the UUID has been
resolved, in this presumably unusual edge case.

fixes #235