systemd: Setup /var/lib/signald
Recent systemd versions (>= 235) can setup /var/lib/SOMEDIR
with appropriate permissions.

So use this to create /var/lib/signald and set the
permissions so that the signald group can read/enter the
directory, but not modify it (create/delete
files/directories).

This is in addition to any other means of setting up this
directory until systemd 235 is common enough.