contacts: add tests
This commit includes adding tests for both Postgres and SQLite for the
following operations:

* Migration from JSON to the database
* Upsert/get/partial update of contacts from the database

The Postgres server is configurable using the
SIGNALD_POSTGRES_TEST_DATABASE environment variable. It must be set to a
Postgres URI that allows the user to log in and create databases.

This commit also updates the CI configuration to run postgres with the
tests.