Use jlink to build minimal docker image
This changes the container image to be much smaller, and not run as root! Note
that a side effect of this change is that your existing volume's permissions
will need to be updated. To help ease the transition, the image will start as
root and run chown on the volume before starting signald as non-root. A future
update (after the next release) will remove this transition mechanism.

For a preview of future behavior or to avoid using the transition mechanism for
whatever other reason, a tag ending in -non-root is also built and will be
through at least 0.18.0